+++ src/index.ts
@@ -1,7 +1,7 @@
 import './intl-polyfill'
 
 export { AirStrings } from './airstrings'
-export type { AirStringsEvents } from './airstrings'
+export type { AirStringsEvents, ExposureEvent } from './airstrings'
 export type { AirStringsConfig } from './airstrings-config'
 export type { AirStringsError, AirStringsErrorCode } from './airstrings-error'
 export type { BundleStore, StoredBundle } from './storage/bundle-store'